+++ packages/volto/src/helpers/Url/Url.js
@@ -141,7 +141,7 @@ export const isCmsUi = memoize((currentPathname) => {
   return settings.nonContentRoutes.reduce(
     (acc, route) =>
       acc ||
-      (settings.publicNonContentRoutes?.indexOf(route) < 0 &&
+      (!settings.publicNonContentRoutes?.includes(route) &&
         new RegExp(route).test(fullPath)),
     false,
   );
